refactor: remove "mouse over displayer" from the state
Use a ref instead, as the value is not used for rendering.
