+++ src/component/reducer/IgnoreActions.ts
@@ -21,7 +21,6 @@ const ignoreActions = new Set<Partial<Action['type']>>([
   'CHANGE_SPECTRUM_VISIBILITY',
   'CHANGE_SPECTRA_VISIBILITY_BY_NUCLEUS',
   'CHANGE_ACTIVE_SPECTRUM',
-  'SET_MOUSE_OVER_DISPLAYER',
   'CHANGE_FLOAT_MOLECULE_POSITION',
   'ALIGN_SPECTRA',
   'TOGGLE_REAL_IMAGINARY_VISIBILITY',
